refactor(api): encapsulate reusable server functions
