+++ src/app/api/payment/route.ts
@@ -6,11 +6,12 @@
  */
 import { PRODUCTS } from '@/data/products';
 import { db } from '@/db';
-import { updateMemberExpiryDate } from '@/db/queries';
 import { memberTable } from '@/db/schema';
 import { env } from '@/env.mjs';
 import { redisClient } from '@/lib/redis';
 import { squareClient } from '@/lib/square';
+import { updateMemberExpiryDate } from '@/server/update-member-expiry-date';
+import { verifyMembershipPayment } from '@/server/verify-membership-payment';
 import { currentUser } from '@clerk/nextjs';
 import { eq } from 'drizzle-orm';
 import type { CreatePaymentLinkRequest } from 'square';
@@ -116,4 +117,14 @@ export async function PUT(request: Request) {
             .where(eq(memberTable.id, reqBody.data.id));
     }
     return Response.json({ success: true });
+}
+
+export async function GET() {
+    const user = await currentUser();
+    if (!user) {
+        return new Response(null, { status: 401 });
+    }
+
+    const { paid } = await verifyMembershipPayment(user.id);
+    return Response.json({ paid });
 }
